Show poster image until playback or manual seek
Previously, when we set the variants of Shaka player, the poster would
disappear. In general, according to MDN, the poster is only shown up
until the first frame has been loaded.

We override this behavior by handling the poster image manually. The
poster should be visible as long playback hasn't started and the user
hasn't manually seeked to a position. (when the video seeks because of
timecode parameter, the poster should still show at first.)